Add section 5 code prog

+++ code_progression/5/ray/ray.h
@@ -11,7 +11,7 @@ typedef struct {
 
 point3_t ray_at(ray_t r, double t);
 
-double hit_sphere(point3_t center, double radius, ray_t r);
+int hit_sphere(point3_t center, double radius, ray_t r);
 
 color_t ray_color(ray_t r);
